Revert patch 8.0.0851
Patch 8.0.0851: 'smartindent' is used even when 'indentexpr' is set
Commit: 69a76feda9e9d308be6b5fc2185286a061dfecd6

This breaks benizi/vim-endwise for `&ft == "sh"`. Probably just need to
upgrade it, but just pushing this up to GitHub for the moment while
diagnosing the issue.

Example breakage, with `▏` standing in for the cursor position.

Starting from:

    #!/bin/sh
    if true▏

With the patch, hitting `<CR>` indents the following `fi` too far:

    #!/bin/sh
    if true
      ▏
      fi

With the patch reverted, it works as expected:

    #!/bin/sh
    if true
      ▏
    fi